Create CommonUtilities project, move Base32 there
