+++ Microsoft.GA4GH.TES.sln
@@ -23,6 +23,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "deploy-tes-on-azure", "src\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "deploy-tes-on-azure.Tests", "src\deploy-tes-on-azure.Tests\deploy-tes-on-azure.Tests.csproj", "{19A5D64F-5252-4FE7-B7CD-49BDC9185A1B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonUtilities", "src\CommonUtilities\CommonUtilities.csproj", "{D863F869-4C87-4C4C-AF6A-A699E7301C83}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonUtilities.Tests", "src\CommonUtilities.Tests\CommonUtilities.Tests.csproj", "{B4E82DE4-16F7-40D5-9619-702A7EECBE92}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +53,14 @@ Global
 		{19A5D64F-5252-4FE7-B7CD-49BDC9185A1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19A5D64F-5252-4FE7-B7CD-49BDC9185A1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{19A5D64F-5252-4FE7-B7CD-49BDC9185A1B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D863F869-4C87-4C4C-AF6A-A699E7301C83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D863F869-4C87-4C4C-AF6A-A699E7301C83}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D863F869-4C87-4C4C-AF6A-A699E7301C83}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D863F869-4C87-4C4C-AF6A-A699E7301C83}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4E82DE4-16F7-40D5-9619-702A7EECBE92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4E82DE4-16F7-40D5-9619-702A7EECBE92}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4E82DE4-16F7-40D5-9619-702A7EECBE92}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4E82DE4-16F7-40D5-9619-702A7EECBE92}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
