+++ src/TesApi.Web/BatchScheduler.BatchPools.cs
@@ -38,7 +38,7 @@ namespace TesApi.Web
 
             // Generate hash of everything that differentiates this group of pools
             var displayName = $"{vmName}:{vmSize}:{isPreemptable}:{registryServer}:{identityResourceId}";
-            var hash = BatchUtils.ConvertToBase32(SHA1.HashData(Encoding.UTF8.GetBytes(displayName))).TrimEnd('='); // This becomes 32 chars
+            var hash = CommonUtilities.Base32.ConvertToBase32(SHA1.HashData(Encoding.UTF8.GetBytes(displayName))).TrimEnd('='); // This becomes 32 chars
 
             // Build a PoolName that is of legal length, while exposing the most important metadata without requiring user to find DisplayName
             // Note that the hash covers all necessary parts to make name unique, so limiting the size of the other parts is not expected to appreciably change the risk of collisions. Those other parts are for convenience
@@ -136,7 +136,7 @@ namespace TesApi.Web
 
                 var uniquifier = new byte[8]; // This always becomes 13 chars when converted to base32 after removing the three '='s at the end. We won't ever decode this, so we don't need the '='s
                 RandomNumberGenerator.Fill(uniquifier);
-                var poolId = $"{key}-{BatchUtils.ConvertToBase32(uniquifier).TrimEnd('=')}"; // embedded '-' is required by GetKeyFromPoolId()
+                var poolId = $"{key}-{CommonUtilities.Base32.ConvertToBase32(uniquifier).TrimEnd('=')}"; // embedded '-' is required by GetKeyFromPoolId()
                 var modelPool = await modelPoolFactory(poolId);
                 modelPool.Metadata ??= new List<BatchModels.MetadataItem>();
                 modelPool.Metadata.Add(new(PoolHostName, this.hostname));
@@ -234,7 +234,7 @@ namespace TesApi.Web
             => batchPools.Add(pool);
 
         private static string GetKeyFromPoolId(string poolId)
-            => poolId[..poolId.LastIndexOf('-')];
+            => poolId is null || !poolId.Contains('-') ? string.Empty : poolId[..poolId.LastIndexOf('-')];
 
         private class BatchPoolEqualityComparer : IEqualityComparer<IBatchPool>
         {
@@ -312,6 +312,12 @@ namespace TesApi.Web
                     : base(new BatchPoolEqualityComparer())
                 {
                     Key = GetKeyForItem(pool);
+
+                    if (string.IsNullOrWhiteSpace(Key))
+                    {
+                        throw new ArgumentException(default, nameof(pool));
+                    }
+
                     if (!Add(pool))
                     {
                         throw new InvalidOperationException();
